+++ bin/sentry.py
@@ -1,9 +1,6 @@
 import getpass
 
 import sentry_sdk
-from sentry_sdk.integrations.atexit import AtexitIntegration
-from sentry_sdk.integrations.dedupe import DedupeIntegration
-from sentry_sdk.integrations.excepthook import ExcepthookIntegration
 
 from oaloader import office_installation, system_info
 
@@ -22,12 +19,6 @@ def setup_sentry(sentry_dsn: str):
     sentry_sdk.init(
         sentry_dsn,
         send_default_pii=True,
-        auto_enabling_integrations=False,
-        integrations=(
-            ExcepthookIntegration(),
-            DedupeIntegration(),
-            AtexitIntegration(),
-        ),
     )
     sentry_sdk.set_user(
         dict(
@@ -36,4 +27,8 @@ def setup_sentry(sentry_dsn: str):
         )
     )
     sentry_sdk.set_context("System Info", system_info())
-    sentry_sdk.set_context("Word Installation", office_installation("word"))
+    office = office_installation("word")
+    sentry_sdk.set_context("Word Installation", office)
+    sentry_sdk.set_tag("office.name", office["name"])
+    sentry_sdk.set_tag("office.version", office["version"])
+    sentry_sdk.set_tag("office.build", office["build"])