+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.4.1] - 2021-08-26
+### Changed
+- `setup_sentry` now `set_tag` about office installation and use default integrations.
+- installer/uninstaller shows version now.
 ### Fixed
 - Fix `fix_app_error` deletes wrong keys.